ml-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
torch==2.6.0
torchvision==0.21.0
Pillow==11.0.0
numpy==2.2.0
opencv-python-headless==4.10.0.84
httpx==0.27.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
